OIBSIP_TASK2/requirements.txt
mysql-connector-python
pyqt5
emoji
plyer
opencv-python
pillow